requirements.txt
asgiref==3.6.0
astroid==2.14.2
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==3.14.0
django-countries==7.5.1
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.9.0
httpie==3.2.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.1
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
phonenumbers==8.13.7
phonenumberslite==8.13.7
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==3.0.0
psycopg2==2.9.5
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pylint==2.16.2
PySocks==1.7.1
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.4
sqlparse==0.4.3
tomlkit==0.11.6
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.19.0
wrapt==1.15.0
yarg==0.1.9
